requirements.txt
asgiref==3.2.3
Django==3.0.3
django-choices==1.7.1
django-extensions==2.2.8
djangorestframework==3.11.0
psycopg2-binary==2.8.4
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
